backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database Drivers
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis and Caching
redis==5.0.1
hiredis==2.3.2

# Neo4j Graph Database
neo4j==5.16.0

# Vector Database
qdrant-client==1.7.3

# Google AI
google-generativeai==0.3.2
google-cloud-aiplatform==1.42.1

# Async Task Queue
celery==5.3.6
flower==2.0.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# NLP and Embeddings
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0